requirements.txt
psycopg2-binary
smbus2
RPi.GPIO
pyserial
pimoroni-bme280
ltr559
enviroplus
requests
